@@ -1,8 +1,7 @@
 package com.basictest;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Unit test for simple App.
